aimodelservice/requirements.txt
kafka-python
Flask
pandas
scikit-learn
numpy